oic: Correct packet type for 'number' is 'float'
It was 'double', but this is not the correct name of Soletta float
packet.

+++ data/scripts/sol-oic-gen.py
@@ -118,7 +118,7 @@ JSON_TO_SOL_JSON = {
     "string": "string",
     "integer": "int",
     "boolean": "boolean",
-    "number": "double"
+    "number": "float"
 }
 
 def props_are_equivalent(p1, p2):
